News_Summarization/requirements.txt
pandas
scikit-learn
nltk
contractions
keras-preprocessing
tensorflow
numpy